Starts of elasticsearch.  Fixed issue when Meer is haned a bad IP.

+++ src/output-plugins/redis.c
@@ -47,7 +47,7 @@ struct _MeerCounters *MeerCounters;
 struct _MeerHealth *MeerHealth;
 
 
-uint32_t redis_batch_count = 0;
+uint16_t redis_batch_count = 0;
 
 char redis_batch[MAX_REDIS_BATCH][10240 + PACKET_BUFFER_SIZE_DEFAULT];
 char redis_batch_key[MAX_REDIS_BATCH][128] = {{ 0 }};
@@ -224,7 +224,7 @@ bool Redis_Writer ( char *command, char *key, char *value, int expire )
 void JSON_To_Redis ( const char *json_string, const char *key )
 {
 
-    int i = 0;
+    uint16_t i = 0;
 
     /* Write request to Redis queue */
